+++ inyoka/middlewares/common.py
@@ -18,6 +18,7 @@ from django.conf import settings
 from django.contrib import messages
 from django.http import HttpResponse
 from django.middleware.common import CommonMiddleware
+from django.utils.functional import cached_property
 from django_hosts.middleware import HostsRequestMiddleware
 
 from inyoka.utils.local import local, local_manager
@@ -27,6 +28,24 @@ from inyoka.utils.timer import StopWatch
 
 class CommonServicesMiddleware(HostsRequestMiddleware, CommonMiddleware):
     """Hook in as first middleware for common tasks."""
+
+    @cached_property
+    def table_control_characters(self):
+        """
+        Returns a translation table which removes all control characters not allowed in XML/HTML.
+        Use it with `str.translate`.
+
+        The characters can be also described by the regex r'[\x00-\x08\x0B-\x0C\x0E-\x1F]'.
+        However, the regex performed slower in a little benchmark comparison.
+        """
+        table = {0x0B: None, 0x0C: None}
+        for i in range(0x0, 0x08 + 1):
+            table[i] = None
+
+        for i in range(0x0E, 0x1F + 1):
+            table[i] = None
+
+        return table
 
     def process_request(self, request):
         # populate the request
@@ -41,6 +60,16 @@ class CommonServicesMiddleware(HostsRequestMiddleware, CommonMiddleware):
         # It's also mentioned in https://datatracker.ietf.org/doc/html/rfc3986#section-7.3
         if '\x00' in request.path:
             return HttpResponse(content='The URL contained NUL characters', status=400, content_type='text/plain')
+
+        # remove control characters in all POST fields
+        # Thus, these will not slip into the DB and will not be displayed to users in HTML/XML
+        # request.POST does not contain file uploads
+        if request.method == "POST":
+            new_post = request.POST.copy()
+            for key, value_list in new_post.lists():
+                new_post.setlist(key, [v.translate(self.table_control_characters) for v in value_list])
+            new_post._mutable = False
+            request.POST = new_post
 
         # IMPORTANT: Since we run some setup-code (mainly locals), this middleware
         # needs to be the first one, hence we manually dispatch to HostsMiddleware