+++ tests/apps/forum/test_views.py
@@ -1904,6 +1904,59 @@ class TestTopicFeed(TestCase):
         response = self.client.get(f'/feeds/topic/{topic.slug}/short/10/', follow=True)
         self.assertEqual(response.status_code, 403)
 
+    def test_submitted_post_with_control_characters(self):
+        """
+        Test that control characters
+          - get stripped from a post and
+          - don't raise an exception upon visiting a feed
+        """
+        registered_group = Group.objects.get(name=settings.INYOKA_REGISTERED_GROUP_NAME)
+        for privilege in ('forum.view_forum', 'forum.add_topic_forum', 'forum.add_reply_forum'):
+            assign_perm(privilege, registered_group, self.forum)
+
+        self.client.login(username='user', password='user')
+        postdata = {
+            'text': 'control characters \x08 \x0f in text',
+            'send': True
+        }
+        self.client.post(f'/topic/{self.topic.slug}/reply/', postdata)
+
+        self.topic.refresh_from_db()
+        # \x08 and x0f should be stripped from the text
+        self.assertEqual(self.topic.last_post.text, 'control characters   in text')
+
+        self.client.get(f'/feeds/topic/{self.topic.slug}/full/50/')
+        self.assertEqual(int(self.topic.post_count()), 2)
+
+    def test_topic_title_control_characters(self):
+        """
+        Test that control characters
+          - get stripped from a topic title and
+          - don't raise an exception upon visiting a feed
+        """
+        self.subforum = Forum.objects.create(name='sub', parent=self.forum)
+
+        registered_group = Group.objects.get(name=settings.INYOKA_REGISTERED_GROUP_NAME)
+        for privilege in ('forum.view_forum', 'forum.add_topic_forum', 'forum.add_reply_forum'):
+            assign_perm(privilege, registered_group, self.forum)
+            assign_perm(privilege, registered_group, self.subforum)
+
+        anonymous_group = Group.objects.get(name=settings.INYOKA_ANONYMOUS_GROUP_NAME)
+        assign_perm('forum.view_forum', anonymous_group, self.subforum)
+
+        self.client.login(username='user', password='user')
+
+        postdata = {
+           'title': 'control characters \x08 \x0f',
+           'ubuntu_distro': constants.get_distro_choices()[2][0],
+           'text': 'newpost text',
+           'send': True
+        }
+        self.client.post('/forum/%s/newtopic/' % self.subforum.slug, postdata)
+
+        response = self.client.get(f'/feeds/forum/{self.subforum.slug}/full/10/')
+        self.assertContains(response, postdata['text'])
+
     def test_content_exact(self):
         response = self.client.get(f'/feeds/topic/{self.topic.slug}/full/50/')